Fixed more compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@11650 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/modes/world_status.hpp
@@ -155,7 +155,7 @@ public:
 
     // ------------------------------------------------------------------------
     /** Returns the current race time. */
-    float   getTime() const      { return m_time; }
+    float   getTime() const      { return (float)m_time; }
     
     // ------------------------------------------------------------------------
     /** Will be called to notify your derived class that the clock,
